+++ mrViewer/lib/mrvWidgets/mrvTimelineSlider.cpp
@@ -109,11 +109,16 @@ namespace mrv
         return _p->thumbnailCreator;
     }
 
+    void TimelineSlider::hideThumbnail_cb( TimelineSlider* t )
+    {
+        t->hideThumbnail();
+    }
+    
     void TimelineSlider::hideThumbnail()
     {
         TLRENDER_P();
         if ( ! p.thumbnailWindow  ) return;
-
+        
         p.thumbnailWindow->hide();
     }
 
@@ -212,7 +217,8 @@ namespace mrv
         {
             if ( p.thumbnailCreator && p.thumbnailRequestId )
                 p.thumbnailCreator->cancelRequests( p.thumbnailRequestId );
-            hideThumbnail();
+            Fl::add_timeout( 0.005, (Fl_Timeout_Handler) hideThumbnail_cb,
+                             this );
             return 1;
         }
         else if ( e == FL_RELEASE )